dff/requirements.txt
df_engine==0.8.1
nltk==3.5
pyyaml==6.0
pycld2==0.41
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
aiohttp>=3.8.0,<3.9.0